Encapsulate implementation details
