--- src/core/components/Circle/Circle.js
@@ -1,7 +1,0 @@
-import "./Circle.css";
-
-import React from "react";
-
-export default function Circle({ color }) {
-  return <div className="Circle" style={{ backgroundColor: color }} />;
-}
