--- src/core/components/FeaturesList/FeaturesList.js
@@ -1,9 +1,0 @@
-import "./FeaturesList.css";
-
-import React from "react";
-
-export default function FeaturesList({ children }) {
-  return (
-    <div className="FeaturesList">{children}</div>
-  );
-}
