+++ src/dashboard/components/FeaturesList/index.js
