+++ src/dashboard/components/Layout/Layout.js
@@ -4,11 +4,17 @@ import React from "react";
 import classNames from "classnames";
 
 import Container from "../../../core/components/Container";
+import Center from "../../../core/components/Center";
+import Page from "../../../core/components/Page";
 
 export default function Layout({ children, className }) {
   return (
-    <Container className={classNames("Layout", className)}>
-      {children}
-    </Container>
+    <Page>
+      <Center>
+        <Container padding="48px 24px"className={classNames("Layout", className)}>
+          {children}
+        </Container>
+      </Center>
+    </Page>
   );
 }
