+++ src/dashboard/components/Title.js
@@ -1,0 +1,12 @@
+import React from "react";
+
+import Container from "../../core/components/Container";
+import Heading from "../../core/components/Heading";
+
+export default function Title() {
+  return (
+    <Container margin="20px 0px 0px">
+      <Heading as="h5">Welcome to Youst!</Heading>
+    </Container>
+  );
+}
